refactor(debug): Extract debug drawer modules to separate file.
Also removed unnecessary context parameters on `initializeDI()` and
`initializeMapLibApp()`
