+++ app/src/release/kotlin/com/omricat/maplibrarian/AppInit.kt
@@ -5,4 +5,4 @@ import android.content.Context
 /**
  * Initialization that should be done for release build variants
  */
-internal fun MapLibraryApp.initializeMapLibApp(context: Context) {}
+internal fun MapLibraryApp.initializeMapLibApp() {}
